requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.1
Django==5.0.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
